resplat/requirements.txt
beartype==0.18.5
colorama==0.4.6
colorspacious==1.1.2
dacite==1.8.1
e3nn==0.5.1
einops==0.8.1
hydra-core==1.3.2
jaxtyping==0.3.2
lpips==0.1.4
matplotlib==3.10.1
moviepy==2.2.1
numpy==1.26.4
opencv_python==4.11.0.86
Pillow==11.1.0
plyfile==1.1
pytorch_lightning==2.4.0
scikit-image==0.24.0
scipy==1.15.2
sk-video==1.1.10
tqdm==4.67.1
wandb==0.17.7
